FCT création d'un score

+++ flappyBirb_serveur/Models/Score.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string? Pseudo { get; set; } = null;
         public string? Date { get; set; } = null;
-        public int TimeInSeconds { get; set; }
+        public float TimeInSeconds { get; set; }
         public int ScoreValue { get; set; }
         public bool IsPublic { get; set; }
     }
